Added Klak Spout for showing at the real dome.
Cloned github.com/keijiro/KlakSpout and copied the assets folder contents into this project's assets folder.  Duplicated the dome camera and created a new game object with Spout component added and attached it to the new camera (IIRC).  You may have to heavily diff this to see how this was done.